requirements.txt
# Core dependencies
python-dotenv
google-generativeai
mcp
# Async handling
anyio>=4.5
